Extract common cluster state and split clusterNode struct
Move 37 protocol-agnostic functions from cluster_legacy.c to the new
cluster_state.c: node accessors, topology queries, node registry,
shard management, and bitmap helpers.

Introduce clusterNodeLegacyData struct for fields specific to the
legacy cluster bus protocol: configEpoch, ping_sent, pong_received,
meet_sent, fail_time, orphaned_time, last_in_ping_gossip, and
fail_reports. These are accessed via the LEGACY_DATA(node) macro.
Common fields (network identity, slot ownership, replication topology,
transport links) remain directly on clusterNode.

Also move verifyClusterNodeId declaration to cluster.h.

Part of the cluster protocol separation refactoring to allow
alternative cluster bus implementations.
