+++ src/Makefile
@@ -464,6 +464,7 @@ ENGINE_SERVER_OBJ = \
     childinfo.o \
     cluster.o \
     cluster_legacy.o \
+    cluster_state.o \
     cluster_migrateslots.o \
     cluster_slot_stats.o \
     commandlog.o \
